agents/philosopher/pyproject.toml
[project]
name = "philosopher-agent"
version = "0.1.0"
description = "Wise Philosopher Agent for Shinox mesh"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0.0",
    "faststream[cli,kafka]>=0.6.5",
]

[tool.uv.sources]
shinox-agent-sdk = { path = "../../shinox-agent-sdk", editable = true }
